// src/types/userTypes.ts
export interface IUserProps {
    nome?: string;
    sobrenome?: string;
    telefone?: string;
    celular?: string;
    numero?: string;
    cep?: string;
    endereco?: string;
    email?: string;
    confirmarEmail?: string;
    senha?: string;
    confirmarSenha?: string;
    nomeUsuario?: string;
  }